Add configurable base image update strategy
Users might use an image as a base image *and* a normal image. If
Lightouse updates the base image, it will overwrite the existing local
image the other container is running. Therefore, Lighthouse will not be
able to resolve the image tag of the running container and fail to
check for its updates.

This patch makes the behaviour configurable.